Fix compatibility with non-DD charts (#3)
* Fix binaries for v0.3.0

* Fix compatibility with non-DD charts

By default this plugin was base64 encoding the values output into the
decoded secrets file.  This was a convenience for DroneDeploy charts
where we could skip putting `| b64encode` in the templates, but breaks
usage with third party charts that are expecting the plaintext value,
and where we don't control the template.

Added an additional protocal handler `gsmb64` that triggers the previous
behavior and can be used with DD charts until they are converted.

The regular `gsm` protocol handler will no longer output base64 encoded
values to the decoded file.